feat: create deno & web version
